:recycle: rename: 블록체인의 모든 기능 및 도메인들을 분리하지 않고 blockchain으로 통합 #45
