--- src/main/java/com/automated/trading/stock/StockManager/blockchain/block/service/BlockService.java
@@ -1,13 +1,0 @@
-package com.automated.trading.stock.StockManager.blockchain.block.service;
-
-import com.automated.trading.stock.StockManager.blockchain.block.controller.dto.GenerateHashDto;
-
-public interface BlockService {
-
-    String generateHash(GenerateHashDto generateHashDto);
-
-    void createBlock(int member_id);
-
-    void deleteBlock(int block_id);
-
-}
