--- src/main/java/com/automated/trading/stock/StockManager/blockchain/data/DataController.java
@@ -1,4 +1,0 @@
-package com.automated.trading.stock.StockManager.blockchain.data;
-
-public class DataController {
-}
