--- src/main/java/com/automated/trading/stock/StockManager/blockchain/data/DataServiceImpl.java
@@ -1,9 +1,0 @@
-package com.automated.trading.stock.StockManager.blockchain.data;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-public class DataServiceImpl implements DataService {
-}
