--- src/main/java/com/automated/trading/stock/StockManager/blockchain/transaction/TransactionController.java
@@ -1,13 +1,0 @@
-package com.automated.trading.stock.StockManager.blockchain.transaction;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TransactionController {
-
-    private final TransactionService transactionService;
-
-    public TransactionController(TransactionService transactionService) {
-        this.transactionService = transactionService;
-    }
-}
